examples/modx: Add ESP8266 support.
The ESP8266 doesn't seem to support truly position-independent code so
we have to adjust a few pointers after loading of the .mpy file. This
requires reading .rela relocations and adding some code to patch those
pointers on import.

Also, the ESP8266 doesn't use MP_PLAT_ALLOC_EXEC but MP_PLAT_COMMIT_EXEC
so we have to dance around with allocation a bit.